mido: Restore better load balancing after boot
As it turned out, allowing background, system-background, foreground,
and top-app to use all 4 cores results in severe jitter and lag whenever
background tasks are running, whether it be on boot, or during waking up
from sleep.

This commit readds pinning background to core 0-1, system-background and
foreground to cores 0-2, and top-app to cores 0-3.

* Use mido oreo based values

Change-Id: Idb76879e75e05b437600f0b689c260c674eb7eea
Signed-off-by: zainarbani <zaintsyariev@gmail.com>